requirements.txt
snap-plugin-lib-py

